wechat_edu_agent/requirements.txt
openai>=1.30.0
pydantic>=2.6.0
python-dotenv>=1.0.1
